requirements-api.txt
# Dépendances spécifiques à l'API pour le conteneur Docker
numpy
requests
python-dotenv
fastapi
uvicorn
gunicorn
tensorflow-cpu